feat: refactor route imports to use route.component; update product IDs and paths in tests and components
