+++ apps/react-router/app/routes/_root.$.tsx
@@ -1,4 +1,6 @@
-import RouteComponent, { links } from '@remix-docs-ja/base/routes/_root.$/route'
+import RouteComponent, {
+  links,
+} from '@remix-docs-ja/base/routes/_root.$/route.component'
 export default RouteComponent
 export { loader } from '@remix-docs-ja/base/routes/_root.$/route.server'
 export { links }
