+++ apps/react-router/app/routes/_root.tsx
@@ -1,4 +1,4 @@
-import RouteComponent from '@remix-docs-ja/base/routes/_root/route'
+import RouteComponent from '@remix-docs-ja/base/routes/_root/route.component'
 export default RouteComponent
 export { loader } from '@remix-docs-ja/base/routes/_root/route.server'
