+++ packages/base/app/features/product/get-domain.test.ts
@@ -10,7 +10,7 @@ describe('getDomain', () => {
   }> = [
     {
       description: 'should add city to localhost',
-      input: { url: 'https://localhost:5173/', productId: 'remix-docs-ja' },
+      input: { url: 'https://localhost:5173/', productId: 'remix' },
       expected: 'https://remix-docs-ja.localhost:5173/',
     },
     {
@@ -26,14 +26,14 @@ describe('getDomain', () => {
     },
     {
       description: 'should add city to hyper-local.app',
-      input: { url: 'https://techtalk.jp/', productId: 'react-router-docs-ja' },
+      input: { url: 'https://techtalk.jp/', productId: 'react-router-v7' },
       expected: 'https://react-router-docs-ja.techtalk.jp/',
     },
     {
       description: 'should change city for hyper-local.app',
       input: {
         url: 'https://remix-docs-ja.techtalk.jp/',
-        productId: 'react-router-docs-ja',
+        productId: 'react-router-v7',
       },
       expected: 'https://react-router-docs-ja.techtalk.jp/',
     },
@@ -45,14 +45,14 @@ describe('getDomain', () => {
     },
     {
       description: 'should add product id to regular domain',
-      input: { url: 'https://example.com/', productId: 'remix-docs-ja' },
+      input: { url: 'https://example.com/', productId: 'remix' },
       expected: 'https://remix-docs-ja.example.com/',
     },
     {
       description: 'should change product id for regular domain',
       input: {
         url: 'https://remix-docs-ja.example.com/',
-        productId: 'react-router-docs-ja',
+        productId: 'react-router-v7',
       },
       expected: 'https://react-router-docs-ja.hostname.com/',
     },
@@ -67,7 +67,7 @@ describe('getDomain', () => {
         'should preserve path and query parameters for special domain',
       input: {
         url: 'https://remix-docs-ja.techtalk.jp/path?param1=value1&param2=value2',
-        productId: 'react-router-docs-ja',
+        productId: 'react-router-v7',
       },
       expected:
         'https://react-router-docs-ja.hyper-local.app/path?param1=value1&param2=value2',
@@ -77,7 +77,7 @@ describe('getDomain', () => {
         'should preserve path, query parameters, and fragment for regular domain',
       input: {
         url: 'https://example.com/page?query=test#section1',
-        productId: 'react-router-docs-ja',
+        productId: 'react-router-v7',
       },
       expected:
         'https://react-router-docs-ja.example.com/page?query=test#section1',
@@ -87,7 +87,7 @@ describe('getDomain', () => {
         'should preserve path, query parameters, and fragment when changing product id for regular domain',
       input: {
         url: 'https://react-router-docs-ja.example.com/page?query=test#section1',
-        productId: 'remix-docs-ja',
+        productId: 'remix',
       },
       expected: 'https://remix-docs-ja.example.com/page?query=test#section1',
     },
@@ -105,7 +105,7 @@ describe('getDomain', () => {
       'https://remix-docs-ja.example.com/path?query=test',
     )
     const originalHref = originalUrl.href
-    getDomain(originalUrl, 'react-router-docs-ja')
+    getDomain(originalUrl, 'react-router-v7')
     expect(originalUrl.href).toBe(originalHref)
   })
 })
