+++ packages/base/app/features/product/products.ts
@@ -4,14 +4,14 @@ export const products = [
     name: 'Remix',
     title: 'Remixドキュメント日本語版',
     url: 'https://remix-docs-ja.techtalk.jp',
-    pagefind: '/pagefind/remix-docs-ja/pagefind.js?url',
+    pagefind: '/pagefind/pagefind.js?url',
   },
   {
     id: 'react-router-v7',
     name: 'React Router',
     title: 'React Router v7 ドキュメント日本語版',
     url: 'https://react-router-docs-ja.techtalk.jp',
-    pagefind: '/pagefind/react-router-docs-ja/pagefind.js?url',
+    pagefind: '/pagefind/pagefind.js?url',
   },
 ] satisfies Product[]
