+++ packages/base/app/routes.ts
@@ -1,3 +1,5 @@
 import { flatRoutes } from '@react-router/fs-routes'
 
-export default flatRoutes()
+export default flatRoutes({
+  ignoredRouteFiles: ['./**/route.tsx'],
+})
