+++ packages/base/app/routes/_root.$/route.component.tsx
