+++ packages/base/app/routes/_root._index/index.tsx
@@ -1,3 +1,3 @@
-import RouteComponent from './route'
+import RouteComponent from './route.component'
 export * from './route.server'
 export default RouteComponent
