+++ packages/base/app/routes/_root._index/route.component.tsx
@@ -1,0 +1,2 @@
+import RouteComponent from '../_root.$/route.component'
+export default RouteComponent
