--- packages/base/app/routes/_root._index/route.tsx
@@ -1,2 +1,0 @@
-import RouteComponent from '../_root.$/route'
-export default RouteComponent
