+++ packages/base/app/routes/_root/index.tsx
@@ -1,3 +1,3 @@
-import RouteComponent from './route'
+import RouteComponent from './route.component'
 export default RouteComponent
 export * from './route.server'
